feat(jsruntime): String.prototype.indexOf
