+++ libs/jsruntime/src/objects/builtins/string/README.md
@@ -1,0 +1,44 @@
+# String
+
+* [ ] [Constructor](https://tc39.es/ecma262/#sec-string-constructor)
+* [x] [String](https://tc39.es/ecma262/#sec-string-constructor)
+* [ ] [String.fromCharCode](https://tc39.es/ecma262/#sec-string.fromcharcode)
+* [ ] [String.fromCodePoint](https://tc39.es/ecma262/#sec-string.fromcodepoint)
+* [ ] [String.raw](https://tc39.es/ecma262/#sec-string.raw)
+* [ ] [String.prototype.at](https://tc39.es/ecma262/#sec-string.prototype.at)
+* [ ] [String.prototype.charAt](https://tc39.es/ecma262/#sec-string.prototype.charat)
+* [ ] [String.prototype.charCodeAt](https://tc39.es/ecma262/#sec-string.prototype.charcodeat)
+* [ ] [String.prototype.codePointAt](https://tc39.es/ecma262/#sec-string.prototype.codepointat)
+* [ ] [String.prototype.concat](https://tc39.es/ecma262/#sec-string.prototype.concat)
+* [x] [String.prototype.constructor](https://tc39.es/ecma262/#sec-string.prototype.constructor)
+* [ ] [String.prototype.endsWith](https://tc39.es/ecma262/#sec-string.prototype.endswith)
+* [ ] [String.prototype.includes](https://tc39.es/ecma262/#sec-string.prototype.includes)
+* [ ] [String.prototype.indexOf](https://tc39.es/ecma262/#sec-string.prototype.indexof)
+* [ ] [String.prototype.isWellFormed](https://tc39.es/ecma262/#sec-string.prototype.iswellformed)
+* [ ] [String.prototype.lastIndexOf](https://tc39.es/ecma262/#sec-string.prototype.lastindexof)
+* [ ] [String.prototype.localeCompare](https://tc39.es/ecma262/#sec-string.prototype.localecompare)
+* [ ] [String.prototype.match](https://tc39.es/ecma262/#sec-string.prototype.match)
+* [ ] [String.prototype.matchAll](https://tc39.es/ecma262/#sec-string.prototype.matchall)
+* [ ] [String.prototype.normalize](https://tc39.es/ecma262/#sec-string.prototype.normalize)
+* [ ] [String.prototype.padEnd](https://tc39.es/ecma262/#sec-string.prototype.padend)
+* [ ] [String.prototype.padStart](https://tc39.es/ecma262/#sec-string.prototype.padstart)
+* [ ] [String.prototype.repeat](https://tc39.es/ecma262/#sec-string.prototype.repeat)
+* [ ] [String.prototype.replace](https://tc39.es/ecma262/#sec-string.prototype.replace)
+* [ ] [String.prototype.replaceAll](https://tc39.es/ecma262/#sec-string.prototype.replaceall)
+* [ ] [String.prototype.search](https://tc39.es/ecma262/#sec-string.prototype.search)
+* [ ] [String.prototype.slice](https://tc39.es/ecma262/#sec-string.prototype.slice)
+* [ ] [String.prototype.splice](https://tc39.es/ecma262/#sec-string.prototype.split)
+* [ ] [String.prototype.startsWith](https://tc39.es/ecma262/#sec-string.prototype.startswith)
+* [ ] [String.prototype.substring](https://tc39.es/ecma262/#sec-string.prototype.substring)
+* [ ] [String.prototype.toLocaleLowerCase](https://tc39.es/ecma262/#sec-string.prototype.tolocalelowercase)
+* [ ] [String.prototype.toLocaleUpperCase](https://tc39.es/ecma262/#sec-string.prototype.tolocaleuppercase)
+* [ ] [String.prototype.toLowerCase](https://tc39.es/ecma262/#sec-string.prototype.tolowercase)
+* [ ] [String.prototype.toString](https://tc39.es/ecma262/#sec-string.prototype.tostring)
+* [ ] [String.prototype.toUpperCase](https://tc39.es/ecma262/#sec-string.prototype.touppercase)
+* [ ] [String.prototype.toWellFormed](https://tc39.es/ecma262/#sec-string.prototype.towellformed)
+* [ ] [String.prototype.trim](https://tc39.es/ecma262/#sec-string.prototype.trim)
+* [ ] [String.prototype.trimEnd](https://tc39.es/ecma262/#sec-string.prototype.trimend)
+* [ ] [String.prototype.trimStart](https://tc39.es/ecma262/#sec-string.prototype.trimstart)
+* [ ] [String.prototype.valueOf](https://tc39.es/ecma262/#sec-string.prototype.valueof)
+* [ ] [String.prototype\[%Symbol.iterator%\]](https://tc39.es/ecma262/#sec-string.prototype-%symbol.iterator%)
+* [ ] [length](https://tc39.es/ecma262/#sec-properties-of-string-instances-length)
